Add connection name and 2 dictionaries
The connection name is used in the 2 dictionaries that are meant to keep track of the connections created.
One, the weight_dictionary has tuples of (pre_id,post_id) as keys and tuples of (name, weight) as values.
The other dictionary has synapse names as keys and tuples of (pre_list, post_list) as values.