Change Readme.md to refer to cloning using git protocol instead of https (#7)
If you followed the instruction to clone using the https URL, you'd then
potentially be faced with a prompt for a password when updating the submodules
if you hadn't already added an ssh key on code.ini.uzh.ch. This is obviously
potentially a bit confusing (see Issue #7).
Therefore to not mix https & git protocols, this commit changes the advice in
the Readme.md to use git from the git-go, so to say, thus avoiding delayed
nasty surprises and confronting people with the need to add an ssh key from
the beginning.